phase3(routing): extract computeModelScore from ModelRouter.scoreModel (+ differential fuzz vs the original)
#12b. The ~680-line capability-scoring arithmetic that decides which model Auto picks lived in the
private ModelRouter.scoreModel and was untestable. Extracted it VERBATIM into a pure, node-tested
common/routing/computeModelScore.ts; ModelRouter.scoreModel is now a thin wrapper that resolves the
impure inputs (getCachedCapabilities, realParamSize from settingsState, this.isVisionCapable, the
freeTierQuotaService lookup, the two globalSettings reads) and delegates.

Proven byte-identical to git HEAD: a whitespace-normalized diff of the extracted scoring body against
HEAD's scoreModel yields ZERO content differences modulo the 5 documented input substitutions (the
injected capabilities/realParamSize/isVisionCapable/getFreeTierRemaining + routingPolicy/localFirstAI
scalars, and 3 em-dash->'--' hygiene edits). A 4-lens adversarial-verify Workflow independently
re-derived this (all lenses: byte-identical; blast-radius clean; oracle is a true independent copy).

Validation: 50k differential fuzz vs a GENERATED oracle (test/common/computeModelScore.oracle.ts =
HEAD's scoreModel body, script-extracted with the same substitutions -> catches any future drift of
the extracted fn from the original) + 11 hand-traced goldens pinning each scoring axis (quality tier,
privacy, code/codebase-question, local-first heavy/light, low-latency, vision, free-tier exhaustion)
+ invariants (auto->0, score>=0, quota-throw swallowed). Removed now-unused imports from modelRouter.
tsgo 0; common suite 798 -> 814 (+16); hygiene clean (0 non-ASCII on added src lines).
